fix: use long.TryParse in NaturalStringComparer to handle large numeric suffixes
Replace int.TryParse with long.TryParse to correctly parse and compare numbers
larger than int.MaxValue (2147483647). This ensures proper numeric sorting for
channel names with large numeric suffixes, maintaining sort transitivity.

Added comprehensive tests:
- Compare_LargeNumericSuffix_SortsNumerically: Tests numbers > int.MaxValue
- Compare_RegularNumbers_SortsNaturally: Tests regular numeric sorting

+++ Application/Channels/NaturalStringComparer.cs
@@ -19,9 +19,9 @@ namespace JSQViewer.Application.Channels
             {
                 if (i >= a.Length) return -1;
                 if (i >= b.Length) return 1;
-                int ai, bi;
-                bool aIsNum = int.TryParse(a[i], out ai);
-                bool bIsNum = int.TryParse(b[i], out bi);
+                long ai, bi;
+                bool aIsNum = long.TryParse(a[i], out ai);
+                bool bIsNum = long.TryParse(b[i], out bi);
                 int cmp = (aIsNum && bIsNum)
                     ? ai.CompareTo(bi)
                     : string.Compare(a[i], b[i], StringComparison.OrdinalIgnoreCase);
